# backend/openapi/api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Platform Backend",
    "description": "Health Platform & Digital Medicine Marketplace backend (modular monolith). **Identity**: registration, authentication, sessions, MFA, social login, and role/permission management. **Onboarding**: applicant onboarding, document review, verification checks, and maker-checker approval for doctors, nutritionists, hospitals, clinics, and pharmacies.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/auth/password/change": {
      "post": {
        "tags": [
          "Password"
        ],
        "summary": "Change the current user's password",
        "operationId": "change_password_api_v1_auth_password_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Current password incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "tags": [
          "Password"
        ],
        "summary": "Request a password reset link",
        "operationId": "forgot_password_api_v1_auth_password_forgot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "tags": [
          "Password"
        ],
        "summary": "Reset a password using a reset token",
        "operationId": "reset_password_api_v1_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Token invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/authenticator/enrol": {
      "post": {
        "tags": [
          "Multi-Factor Authentication"
        ],
        "summary": "Start authenticator-app (TOTP) enrolment",
        "operationId": "enrol_authenticator_api_v1_auth_mfa_authenticator_enrol_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_EnrolAuthenticatorResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/authenticator/confirm": {
      "post": {
        "tags": [
          "Multi-Factor Authentication"
        ],
        "summary": "Confirm authenticator enrolment with a TOTP code",
        "operationId": "confirm_authenticator_api_v1_auth_mfa_authenticator_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmAuthenticatorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_ConfirmAuthenticatorResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Enrolment expired or already confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "Multi-Factor Authentication"
        ],
        "summary": "Complete login by verifying an MFA challenge",
        "description": "method is one of AUTHENTICATOR | SMS | EMAIL | RECOVERY_CODE.",
        "operationId": "verify_mfa_api_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect code or expired challenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/recovery-codes/regenerate": {
      "post": {
        "tags": [
          "Multi-Factor Authentication"
        ],
        "summary": "Regenerate recovery codes (invalidates all unused prior codes)",
        "operationId": "regenerate_recovery_codes_api_v1_auth_mfa_recovery_codes_regenerate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Password confirmation required/incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/{method_id}": {
      "delete": {
        "tags": [
          "Multi-Factor Authentication"
        ],
        "summary": "Disable an MFA method",
        "operationId": "disable_mfa_api_v1_auth_mfa__method_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "method_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableMfaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "401": {
            "description": "Password confirmation required/incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Cannot disable the last mandatory MFA method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "List active sessions",
        "operationId": "list_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_SessionOut__"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "Sessions"
        ],
        "summary": "Revoke a session",
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword"
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "ChangePasswordRequest"
      },
...
      "ConfirmAuthenticatorRequest": {
        "properties": {
          "enrolmentId": {
            "type": "string",
            "format": "uuid",
            "title": "Enrolmentid"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "enrolmentId",
          "code"
        ],
        "title": "ConfirmAuthenticatorRequest"
      },
      "ConfirmAuthenticatorResponse": {
        "properties": {
          "methodId": {
            "type": "string",
            "format": "uuid",
            "title": "Methodid"
          },
          "recoveryCodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recoverycodes"
          }
        },
        "type": "object",
        "required": [
          "methodId",
          "recoveryCodes"
        ],
        "title": "ConfirmAuthenticatorResponse"
      },
...
      "DisableMfaRequest": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword"
        ],
        "title": "DisableMfaRequest"
      },
...
      "EnrolAuthenticatorResponse": {
        "properties": {
          "enrolmentId": {
            "type": "string",
            "format": "uuid",
            "title": "Enrolmentid"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "otpauthUri": {
            "type": "string",
            "title": "Otpauthuri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "enrolmentId",
          "secret",
          "otpauthUri",
          "expiresAt"
        ],
        "title": "EnrolAuthenticatorResponse"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "token",
          "newPassword"
        ],
        "title": "ResetPasswordRequest"
      },
...
      "SessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "deviceName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devicename"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "ipAddressMasked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipaddressmasked"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "lastUsedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastusedat"
          },
          "current": {
            "type": "boolean",
            "title": "Current"
          }
        },
        "type": "object",
        "required": [
          "id",
          "deviceName",
          "platform",
          "ipAddressMasked",
          "createdAt",
          "lastUsedAt",
          "current"
        ],
        "title": "SessionOut"
      },
...
      "SuccessEnvelope_ConfirmAuthenticatorResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/ConfirmAuthenticatorResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[ConfirmAuthenticatorResponse]"
      },
      "SuccessEnvelope_EnrolAuthenticatorResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/EnrolAuthenticatorResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[EnrolAuthenticatorResponse]"
      },
...
      "SuccessEnvelope_list_SessionOut__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[list[SessionOut]]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyMfaRequest": {
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "title": "Challengeid"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "challengeId",
          "method",
          "code"
        ],
        "title": "VerifyMfaRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token issued by /auth/login or /auth/refresh",
        "scheme": "bearer"
      }
    }
  },
...
}